chg: Reactoring in separate files
