+++ signalr-go-server/pkg/signalr/groupmanager.go
@@ -1,0 +1,20 @@
+package signalr
+
+type GroupManager interface {
+	AddToGroup(groupName string, connectionID string)
+	RemoveFromGroup(groupName string, connectionID string)
+}
+
+type defaultGroupManager struct {
+	lifetimeManager HubLifetimeManager
+}
+
+func (d *defaultGroupManager) AddToGroup(groupName string, connectionID string) {
+	d.lifetimeManager.AddToGroup(groupName, connectionID)
+}
+
+func (d *defaultGroupManager) RemoveFromGroup(groupName string, connectionID string) {
+	d.lifetimeManager.RemoveFromGroup(groupName, connectionID)
+}
+
+
